gpio: sunxi: fix empty string process for legacy sunxi_name_to_gpio
The legacy version of sunxi_name_to_gpio function (which is used for
SPL) currently doesn't deal with empty string well, and a empty
string passed to it will return PA0 instead.

This makes eMMC booting fail on H6 when rebooting, which is because the
eMMC card detect pin is considered as PA0.

Fix this function by check empty string and then return -1.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>